all backend

+++ backend/src/main/java/PBL3/backend/service/GoiDichVuService.java
@@ -1,6 +1,7 @@
 package PBL3.backend.service;
 
-import PBL3.backend.dto.GoiDichVuDTO;
+import PBL3.backend.dto.request.GoiDichVuRequest;
+import PBL3.backend.exception.ResourceNotFoundException;
 import PBL3.backend.model.GoiDichVu;
 import PBL3.backend.repository.GoiDichVuRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,27 +24,30 @@ public class GoiDichVuService {
         return goiDichVuRepository.findById(id);
     }
 
-    public GoiDichVu taoGoiDichVu(GoiDichVuDTO goiDichVuDTO) {
+    public GoiDichVu taoGoiDichVu(GoiDichVuRequest goiDichVuRequest) {
         GoiDichVu goiDichVu = new GoiDichVu();
-        goiDichVu.setTenGoi(goiDichVuDTO.getTenGoi());
-        goiDichVu.setGia(goiDichVuDTO.getGia()); // Already BigDecimal in DTO
-        goiDichVu.setMoTa(goiDichVuDTO.getMoTa());
-        goiDichVu.setThoiGian(goiDichVuDTO.getThoiGian());
+        goiDichVu.setTenGoi(goiDichVuRequest.getTenGoi());
+        goiDichVu.setGia(goiDichVuRequest.getGia());
+        goiDichVu.setMoTa(goiDichVuRequest.getMoTa());
+        goiDichVu.setThoiGian(goiDichVuRequest.getThoiGian());
         return goiDichVuRepository.save(goiDichVu);
     }
 
-    public GoiDichVu capNhatGoiDichVu(int id, GoiDichVuDTO goiDichVuDTO) {
+    public GoiDichVu capNhatGoiDichVu(int id, GoiDichVuRequest goiDichVuRequest) {
         GoiDichVu goiDichVu = goiDichVuRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Không tìm thấy gói dịch vụ với ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("GoiDichVu", "id", id));
         
-        goiDichVu.setTenGoi(goiDichVuDTO.getTenGoi());
-        goiDichVu.setGia(goiDichVuDTO.getGia()); // Already BigDecimal in DTO
-        goiDichVu.setMoTa(goiDichVuDTO.getMoTa());
-        goiDichVu.setThoiGian(goiDichVuDTO.getThoiGian());
+        goiDichVu.setTenGoi(goiDichVuRequest.getTenGoi());
+        goiDichVu.setGia(goiDichVuRequest.getGia());
+        goiDichVu.setMoTa(goiDichVuRequest.getMoTa());
+        goiDichVu.setThoiGian(goiDichVuRequest.getThoiGian());
         return goiDichVuRepository.save(goiDichVu);
     }
 
     public void xoaGoiDichVu(int id) {
+        if (!goiDichVuRepository.existsById(id)) {
+            throw new ResourceNotFoundException("GoiDichVu", "id", id);
+        }
         goiDichVuRepository.deleteById(id);
     }
 }
